Fixed pattern wallpapers not showing in other accounts

+++ TelegramUI/FetchVideoMediaResource.swift
@@ -143,9 +143,9 @@ public func fetchVideoLibraryMediaResource(postbox: Postbox, resource: VideoLibr
                         if let result = next as? TGMediaVideoConversionResult {
                             var value = stat()
                             if stat(result.fileURL.path, &value) == 0 {
-                                if config.remuxToFMp4 {
+                                if false && config.remuxToFMp4 {
                                     let tempFile = TempBox.shared.tempFile(fileName: "video.mp4")
-                                    if false && FFMpegRemuxer.remux(result.fileURL.path, to: tempFile.path) {
+                                    if FFMpegRemuxer.remux(result.fileURL.path, to: tempFile.path) {
                                         let _ = try? FileManager.default.removeItem(atPath: result.fileURL.path)
                                         subscriber.putNext(.moveTempFile(file: tempFile))
                                     } else {
@@ -228,9 +228,9 @@ func fetchLocalFileVideoMediaResource(postbox: Postbox, resource: LocalFileVideo
                 if let result = next as? TGMediaVideoConversionResult {
                     var value = stat()
                     if stat(result.fileURL.path, &value) == 0 {
-                        if config.remuxToFMp4 {
+                        if false && config.remuxToFMp4 {
                             let tempFile = TempBox.shared.tempFile(fileName: "video.mp4")
-                            if false && FFMpegRemuxer.remux(result.fileURL.path, to: tempFile.path) {
+                            if FFMpegRemuxer.remux(result.fileURL.path, to: tempFile.path) {
                                 let _ = try? FileManager.default.removeItem(atPath: result.fileURL.path)
                                 subscriber.putNext(.moveTempFile(file: tempFile))
                             } else {
